# openapi.yaml
openapi: 3.1.0
info:
  title: VoiceBlender API
  description: >
    VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio
    mixing, real-time speech-to-text, text-to-speech, AI agent integration,
    recording, and webhook-based event delivery.


    **Configuration environment variables:**
    `INSTANCE_ID` (auto-generated UUID; included in all API response bodies and
    webhook events), `HTTP_ADDR` (:8080), `SIP_BIND_IP`, `SIP_PORT`, `SIP_HOST`,
    `ICE_SERVERS`, `RECORDING_DIR`, `LOG_LEVEL`, `WEBHOOK_URL`,
    `ELEVENLABS_API_KEY`, `VAPI_API_KEY` (VAPI agent provider API key),
    `S3_BUCKET`, `S3_REGION`, `S3_ENDPOINT`, `S3_PREFIX`.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /rooms:
    post:
      operationId: createRoom
      summary: Create a room
      tags: [Rooms]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '409':
          description: Room ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listRooms
      summary: List all rooms with participants
      tags: [Rooms]
      responses:
        '200':
          description: Array of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
  /rooms/{id}:
    parameters:
      - $ref: '#/components/parameters/RoomId'

    get:
      operationId: getRoom
      summary: Get a room with participants
      tags: [Rooms]
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          $ref: '#/components/responses/RoomNotFound'

    delete:
      operationId: deleteRoom
      summary: Delete a room
      tags: [Rooms]
      responses:
        '200':
          description: Room deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: deleted
        '404':
          $ref: '#/components/responses/RoomNotFound'
  /rooms/{id}/legs:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: addLegToRoom
      summary: Add or move a leg to a room
      description: >
        Add a leg to a room (auto-creates room if it doesn't exist).
        If the leg is already in a different room, it is atomically moved
        to the target room. A ringing inbound SIP leg is automatically
        answered before being added.
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddLegRequest'
      responses:
        '200':
          description: Leg added or moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: added
        '400':
          description: Invalid JSON, leg not found, or leg not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /rooms/{id}/legs/{legID}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
      - name: legID
        in: path
        required: true
        schema:
          type: string
        description: Leg ID
    put:
      operationId: moveLegToRoom
      summary: Move a leg to a different room
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to_room_id:
                  type: string
                  description: Destination room ID
              required: [to_room_id]
      responses:
        '200':
          description: Leg moved
          content:
            application/json:
              schema:
                type: object
                properties:
                  instance_id:
                    type: string
                    description: Instance identifier
                  status:
                    type: string
                    example: moved
                  from:
                    type: string
                    description: Source room ID
                  to:
                    type: string
                    description: Destination room ID
                required: [status, from, to]
        '400':
          description: Invalid JSON, leg not found, or move not possible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: removeLegFromRoom
      summary: Remove a leg from a room
      tags: [Rooms]
      responses:
        '200':
          description: Leg removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: removed
        '400':
          description: Room or leg not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
...
    RoomId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Room ID
...
  responses:
...
    RoomNotFound:
      description: Room not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Leg:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        leg_id:
          type: string
          description: Unique leg identifier (UUID)
        type:
          type: string
          enum: [sip_inbound, sip_outbound, webrtc]
        state:
          type: string
          enum: [ringing, early_media, connected, held, hung_up]
        room_id:
          type: string
          description: Room ID if the leg is in a room, empty otherwise
        muted:
          type: boolean
          description: Whether the leg is muted
        held:
          type: boolean
          description: Whether the call is on hold (SIP legs only)
        sip_headers:
          type: object
          additionalProperties:
            type: string
          description: 'X-* headers from the inbound INVITE. Only present on sip_inbound legs.'
      required: [leg_id, type, state]
    Room:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        id:
          type: string
          description: Room identifier
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Leg'
      required: [id, participants]
...
    Error:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        error:
          type: string
          description: Error message
      required: [error]
    StatusResponse:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        status:
          type: string
      required: [status]

    # ── Request bodies ──────────────────────────────────────────────────
...
    RoomCreateRequest:
      type: object
      properties:
        id:
          type: string
          description: Custom room ID (auto-generated UUID if omitted)
        webhook_url:
          type: string
          format: uri
          description: Route all events for this room exclusively to this URL instead of global webhooks.
        webhook_secret:
          type: string
          description: HMAC-SHA256 signing secret for the per-room webhook.
    AddLegRequest:
      type: object
      properties:
        leg_id:
          type: string
          description: ID of the leg to add
      required: [leg_id]

    # ── Response bodies ─────────────────────────────────────────────────
...